Serial: Improve failure behaviour
Previously, the SerialHandler tried to open the port first,
and then do everything else. It also did not check if the
port was correctly opened.
This meant that other components were not properly fetched
and that each Update tried to access the failed port,
generating an error each time.

Move the component fetching before the opening, allowing
them to be properly set an opening up the possibility to
try again later.
Check that the port is open before doing anything.