Make buttons handle timers when collided
Beforehand there was only one button and we only wanted
to know if the player was colliding with it or not, so
this could be handled by the player controller which
already existed.

However, now we want much more from the buttons :
they are two, which need to control two independent
timers and make visual changes to the scene to show
its state.

Introduce a ButtonController that takes care of the
collision detection and controls the newly introduced
timers, as well as a wire to show the activation.

Lighten the PlayerMovement script now that it does
not need to handle collisions or serial.

Introduce a Button Prefab containing the three elements
expected by the script : the button, a wire and a timer.