Serial: Only set serial port config in the inspector
Currently the serial port configuration options, the port name and
the baudrate, are set in both the SerialHandler class and
the Unity inspector.
This can be confusing and is not really necessary : remove
the default values from the code, leaving the inspector as the only place
to update them.

Also : the default serial port on Windows is `COM1`, which is what is in
our SerialHandler class set as well, but was `COM0` in the inspector.
Change it for the "correct" default of `COM1`, like it was in the code.