Protocol: rework the river state messages
The current river state protocol is as follows :
 - "dry" changes the river to the dry state
 - "wet" changes the river to the dry state

This is easy and works, but contrasts with what the students are asked
to do in later parts of the exercise.

Instead, change it to one message with one argument :
 - "river" is the only implemented header,
   - It takes exactly one argument :
     - "0" changes the river to the dry state,
     - "1" changes the river to the wet state,

This has the following benefits :
 - Illustrates how to use/check for arguments
 - Illustrates that numbers are currently strings
 - Is similar to what is asked in question 2.c : a new header with
   a number as an argument
 - Allows unifying and simplifying instructions for the advanced exercise.

Implement this in both the Arduino and Unity code.