Serial: gracefully handle disconnection and opening failure
The SerialPort object provided by .NET is quite limited : it does not
provide a way to check if the serial device is still present before trying
to access it, and will not close a device failing to communicate.
With the current script, this means that the we will just be spammed with
exceptions in case of a deconnection.

We also only try to connect once and will never try again, meaning that
a deconnection or failed initial connection would force a restart.

Implement a function to handle opening of the serial device, catching
exceptions in case of failure so we can retry later.
Update the port name, so it can be updated from the live inspector.
Rate limit it so we can call it in `Update()` without too much worry
or warning spam.

Now that we can retry opening a serial connection, catch exceptions
when reading or writing to it and manually close the connection in those
cases.
This allows the serial device to be plugged back in, correct the port
in the inspector, or fix other issues, without Unity freezing, spamming
errors, or needing to restart the game.