SafeSerial: Introduce a wrapper class sanitizing SerialPort
The .NET SerialPort class correctly produces exceptions when there are
issues with the serial connection, but leaves everything up to the user.

Wrap the class to catch exceptions related to IO and connection issues,
as those are not the focus of what students are trying to learn (the very
basics of serial communication).
Usage exceptions are still bubbled up, as they are relevant to learning
about what should be done.

When catching an exception, forcefully disconnect the SerialPort so we can
try to reconnect in the main SerialHandler, update the port name or even
disconnect/reconnect the Arduino, all without having to restart the game.

If the connection is not opened, warn the user but return a default value.

Return values are added when relevant, trying to follow either default
values in the Mono implementation or what would make the most sense
to receive/handle.