Serial: Use SafeSerial in SerialHandler
Replace the .NET SerialPort with the SafeSerial wrapper.
This allows removing the exception handling and making the core of
the serial communication much clearer and obvious.

This allows students to use an almost identical API (with some returns
being different) to the standard function, but without having to worry
about disconnections, unplugged boards, etc.